Enum poke_escape_server::error::ErrorKind

pub enum ErrorKind {
    IoNotFound,
    IoPermissionDenied,
    IoConnectionRefused,
    IoConnectionReset,
    IoConnectionAborted,
    IoNotConnected,
    IoAddrInUse,
    IoAddrNotAvailable,
    IoBrokenPipe,
    IoAlreadyExists,
    IoWouldBlock,
    IoInvalidInput,
    IoInvalidData,
    IoTimedOut,
    IoWriteZero,
    IoInterrupted,
    IoOther,
    IoUnexpectedEof,
    FormatNotSupported,
    FieldNotExists(String),
    NotParsable(String),
    NoVersionSupplied,
    VersionNotParsable(String),
    PoolToSmall,
    PoolSendError(bool),
    Other(String),
    Unknown(String),
}

ErrorKind represents the internal Error type of the Error

Variants
IoNotFound

Io Not Found error, transformed from std::io::ErrorKind::NotFound

IoPermissionDenied

Io Perssion Denied error, transformed from `std::io::ErrorKind::PermissionDenied

IoConnectionRefused

Io Connection Refused error, transformed from std::io::ErrorKind::ConnectionRefused

IoConnectionReset

Io ConnectionReset error, transformed from std::io::ErrorKind::ConnectionReset

IoConnectionAborted

Io ConnectionAborted error, transformed from std::io::ErrorKind::ConnectionAborted

IoNotConnected

Io Not Connected error, transformed from std::io::ErrorKind::NotConnected

IoAddrInUse

IO Addr In Use error, transformed from std::io::ErrorKind::AddrInUse

IoAddrNotAvailable

Io Addr Not Available error, transformed from std::io::ErrorKind::AddrNotAvailable

IoBrokenPipe

Io Broken Pipe error, transformed from std::io::ErrorKind::BrokenPipe

IoAlreadyExists

IO Already Exists error, transformed from std::io::ErrorKind::AlreadyExists

IoWouldBlock

IO Would Block error, transformed from std::io::ErrorKind::WouldBlock

IoInvalidInput

IO Invalid Input error, transformed from std::io::ErrorKind::InvalidInput

IoInvalidData

IO Invalid Data error, transformed from std::io::ErrorKind::InvalidData

IoTimedOut

IO Timed Out error, transformed from std::io::ErrorKind::TimedOut

IoWriteZero

IO Write Zero error, transformed from std::io::ErrorKind::WriteZero

IoInterrupted

IO Interrupted error, transformed from std::io::ErrorKind::Interrupted

IoOther

IO Other error, transformed from std::io::ErrorKind::Other

IoUnexpectedEof

IO Unexpected EOF error, transformed from `std::io::ErrorKind::UnexpectedEof

FormatNotSupported

Format Not Supported, raised when the format defined in config.toml is not supported

FieldNotExists(String)

Field Not Exists, raised when a important field is missing in config also raised if file is of the wrong type

NotParsable(String)

Not Parsable error, raised when serde reports an error

NoVersionSupplied

No Version Supplied error, used if the version of the client is none

VersionNotParsable(String)

Version Not Parsable error, used if the version cannot be parsed

PoolToSmall

Pool To Small is returned when the Threapool is to small to be created

PoolSendError(bool)

Pool Send Error is a send error in a channel for the ThreadPool contains true if it wasn't a terminate instruction

Other(String)

Other error, used for string to error conversion

Unknown(String)

Unknown error, used for default in match statements

Methods

impl ErrorKind

fn convert_to_string(&self) -> String

represents the type as String

pub fn error_string(&self) -> String

returns the name of the field of the enum as String